requirements.txt
langchain-core>=0.2.38
langchain-community>=0.2.16
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.3
sentence-transformers
transformers
torch
faiss-cpu
pypdf2
python-docx
psutil